web/requirements.txt
python-jenkins
hvac
sanic==0.7.0
websockets==5.0.1